+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 
 - API for BiologicalRelationships -`api/v1/biological_relationships?extend[]=biological_property` [#3525]
+- Organization to Attribution slice in Radial annotator [#3529]
 
 ### Fixed
 
@@ -22,6 +23,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#3525]: https://github.com/SpeciesFileGroup/taxonworks/issues/3525
 [#3526]: https://github.com/SpeciesFileGroup/taxonworks/issues/3526
 [#3528]: https://github.com/SpeciesFileGroup/taxonworks/issues/3528
+[#3529]: https://github.com/SpeciesFileGroup/taxonworks/issues/3529
 
 ## [0.34.2] - 2023-08-16